Simplify and move PF5SpacedListItem (DetailField) to widgets
